platforms/raspberry_pi/requirements.txt
numpy>=1.20.0
scipy>=1.7.0
pyserial>=3.5
smbus2>=0.4.1
RPi.GPIO>=0.7.0
adafruit-circuitpython-mpu6050>=1.1.0